Update Professor.lua
The addon told me that I had "Found 1/0 (-1 Missing)" in the rare section of the Legion archeologic artifacts.
So I realized, that the quest rewards are counted as rare artifacts and are also missing in the database.

For the Shard of Sciallax you can actually choose from 6 different rewards. I arbitrarilly selected the itemID of one of them.